+++ guillotina/tests/fixtures.py
@@ -40,10 +40,11 @@ annotations = {
 }
 
 def base_settings_configurator(settings):
-    settings["load_utilities"]['catalog'] = {
-        "provides": "guillotina.interfaces.ICatalogUtility",
-        "factory": "guillotina.catalog.catalog.DefaultSearchUtility"
-    }
+    if 'catalog' not in settings["load_utilities"]:
+        settings["load_utilities"]['catalog'] = {
+            "provides": "guillotina.interfaces.ICatalogUtility",
+            "factory": "guillotina.catalog.catalog.DefaultSearchUtility"
+        }
 
 
 testing.configure_with(base_settings_configurator)